% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_data.R
\name{check_entered_data}
\alias{check_entered_data}
\alias{make_paramedic_tibbles}
\alias{make_paramedic_stan_data}
\title{Prepare Data for Paramedic}
\usage{
check_entered_data(
  W,
  V,
  X,
  inits_lst,
  sigma_beta,
  sigma_Sigma,
  alpha_sigma,
  kappa_sigma
)

make_paramedic_tibbles(
  W,
  V,
  X,
  inits_lst,
  sigma_beta,
  sigma_Sigma,
  alpha_sigma,
  kappa_sigma
)

make_paramedic_stan_data(
  W_mat,
  V_mat,
  X_mat,
  inits_lst,
  sigma_beta,
  sigma_Sigma,
  alpha_sigma,
  kappa_sigma,
  n_chains,
  centered = FALSE
)
}
\arguments{
\item{W}{The relative abundance data, e.g., from broad range 16S sequencing with "universal" primers. Expects data (e.g., matrix, data.frame, tibble) with sample identifiers in the first column. Sample identifiers must be the same between W and V, and the column must have the same name in W and V.}

\item{V}{The absolute abundance data, e.g., from taxon-specific absolute primers. Expects data (e.g., matrix, data.frame, tibble) with sample identifiers in the first column. Sample identifiers must be the same between W and V, and the column must have the same name in W and V.}

\item{X}{The covariate data. Expects data (e.g., matrix, data.frame, tibble) with sample identifiers in the first column. Sample identifiers must be the same between W, V, and X, and the column must have the same name in W, V, and X. If X only consists of the subject identifiers, then no covariates are used.}

\item{inits_lst}{An optional list of initial values of the parameters. Must be a named list; see \code{\link[rstan]{stan}}.}

\item{sigma_beta}{Hyperparameter specifying the prior variance on \eqn{\beta_0}. Defaults to \eqn{\sqrt{50}}.}

\item{sigma_Sigma}{Hyperparameter specifying the prior variance on \eqn{\Sigma}. Defaults to \eqn{\sqrt{50}}.}

\item{alpha_sigma}{Hyperparameter specifying the shape parameter of the prior distribution on \eqn{\sigma_e}. Defaults to 2.}

\item{kappa_sigma}{Hyperparameter specifying the scale parameter of the prior distribution on \eqn{\sigma_e}. Defaults to 1.}

\item{W_mat}{the pre-processed matrix W}

\item{V_mat}{the pre-processed matrix V}

\item{X_mat}{the pre-processed matrix X}
}
\description{
Functions in \code{paramedic} expect data (W, V, and X) in matrix, data.frame, or tibble form, 
with the first column being sample identifiers. Sample identifiers must be the same 
between W, V, and X and the column must have the same name in W and V. This function
checks the incoming data, transforms it into the data list that is passed to the
Stan functions, and computes any initial values of model parameters.
}
\section{Functions}{
\itemize{
\item \code{make_paramedic_tibbles}: Make tibbles from entered data

\item \code{make_paramedic_stan_data}: Make data and initial values lists to pass to stan
}}

