% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_16S_data.R
\docType{data}
\name{example_16S_data}
\alias{example_16S_data}
\title{Example relative abundance data}
\format{
`example_16S_data`: a tibble with 20 rows and 434 columns, where the first column is the
subject identifiers, and each subsequent column is
the relative abundance of a particular taxon in a given sample (row).
}
\source{
Inspired by data collected for Project 2 of the UW STICRC \url{http://depts.washington.edu/cfas/research/sexually-transmitted-infections-cooperative-research-center/projects/establishing-microbial-and-biochemical-thresholds-for-development-and-persistence-of-bacterial-vaginosis/}.
}
\usage{
example_16S_data
}
\description{
A dataset inspired by data collected by the University of Washington (UW) 
Sexually Transmitted Infections Cooperative Research Center (STICRC). Relative abundance, as measured by 
broad range 16S sequencing with "universal" primers, is available on 20 samples and 433 taxa, and is  
is provided in `example_16S_data`.
}
\keyword{datasets}
