% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_qPCR_data.R
\docType{data}
\name{example_qPCR_data}
\alias{example_qPCR_data}
\title{Example abolute abundance data}
\format{`example_qPCR_data`: a matrix with 20 rows and 7 columns, where each column is the absolute abundance
of a particular taxon in a given sample (row).}
\source{
Inspired by data collected for Project 2 of the UW STICRC \url{http://depts.washington.edu/cfas/research/sexually-transmitted-infections-cooperative-research-center/projects/establishing-microbial-and-biochemical-thresholds-for-development-and-persistence-of-bacterial-vaginosis/}.
}
\usage{
example_qPCR_data
}
\description{
A dataset inspired by data collected by the University of Washington (UW) 
Sexually Transmitted Infections Cooperative Research Center (STICRC). Absolute abundance, 
as measured by taxon-specific absolute primers,
is available on 20 samples but only 7 taxa, and is provided in `example_qPCR_data`.
}
\keyword{datasets}
