% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_wald_interval.R
\name{gen_wald_interval}
\alias{gen_wald_interval}
\title{Generate a Wald-type prediction interval for qPCR}
\usage{
gen_wald_interval(mu, sd, alpha = 0.05, truncate = TRUE)
}
\arguments{
\item{mu}{the estimated concentrations (a vector of length N times q).}

\item{sd}{the estimated standard deviation of the estimated concentrations (a vector of length N times q).}

\item{alpha}{the desired level (defaults to 0.05, corresponding to a 95\% interval).}

\item{truncate}{truncate negative lower limits at zero (defaults to \code{TRUE})}
}
\value{
A (1 - \eqn{\alpha})x100\% Wald-type prediction interval for each qPCR
}
\description{
Return a Wald-type prediction interval for qPCR values given the estimated concentrations.
}
\examples{
## load the package, read in example data
library("paramedic")
data(simple_example_data)

## process the example data
q <- 3
q_obs <- 2
processed_data <- process_data(full_data = simple_example_data, rel_inds = 1:q,
abs_inds = (q + 1):(q + 1 + q_obs),
abs_plus_rel_inds = 1:q_obs,
regex_thr = NA, regex_abs = "", llod = 0,
m_min = 1000, div_num = 1)
## run paramedic (with a *very* small number of iterations, for illustration only)
mod <- paramedic::run_paramedic(W = processed_data$relative, V = processed_data$absolute,
stan_model = "src/stan_files/variable_efficiency.stan", n_iter = 30, n_burnin = 25, n_chains = 1, stan_seed = 4747,
params_to_save = c("mu", "Sigma", "beta", "e"))
## get model summary
mod_summ <- rstan::summary(mod, probs = c(0.025, 0.975))$summary
## get samples
mod_samps <- rstan::extract(mod)
## extract relevant summaries
summs <- extract_posterior_summaries(stan_mod = mod_summ, stan_samps = mod_samps, taxa_of_interest = 1:3,
mult_num = 1, level = 0.95, interval_type = "wald")

}
