% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/no_efficiency.R
\name{no_efficiency}
\alias{no_efficiency}
\title{Run a Stan algorithm to estimate concentrations from combined absolute and relative abundance data without modeling efficiency}
\usage{
no_efficiency(
  W,
  V,
  X = V[, 1, drop = FALSE],
  n_iter = 10500,
  n_burnin = 10000,
  n_chains = 4,
  stan_seed = 4747,
  inits_lst = NULL,
  sigma_beta = sqrt(50),
  sigma_Sigma = sqrt(50),
  ...
)
}
\arguments{
\item{W}{The relative abundance data, e.g., from broad range 16S sequencing with "universal" primers. Expects data (e.g., matrix, data.frame, tibble) with sample identifiers in the first column. Sample identifiers must be the same between W and V, and the column must have the same name in W and V.}

\item{V}{The absolute abundance data, e.g., from taxon-specific absolute primers. Expects data (e.g., matrix, data.frame, tibble) with sample identifiers in the first column. Sample identifiers must be the same between W and V, and the column must have the same name in W and V.}

\item{X}{The covariate data. Expects data (e.g., matrix, data.frame, tibble) with sample identifiers in the first column. Sample identifiers must be the same between W, V, and X, and the column must have the same name in W, V, and X. If X only consists of the subject identifiers, then no covariates are used.}

\item{n_iter}{The total number of iterations per chain to be run by the Stan algorithm. Defaults to 10500.}

\item{n_burnin}{The total number of warmups per chain to be run by the Stan algorithm. Defaults to 10000.}

\item{n_chains}{The total number of chains to be run by the Stan algorithm. Defaults to 4.}

\item{stan_seed}{The random number seed to initialize.}

\item{inits_lst}{An optional list of initial values of the parameters. Must be a named list; see \code{\link[rstan]{stan}}.}

\item{sigma_beta}{Hyperparameter specifying the prior variance on \code{beta_0}. Defaults to \code{\sqrt{50}}.}

\item{sigma_Sigma}{Hyperparameter specifying the prior variance on \code{\Sigma}. Defaults to \code{\sqrt{50}}.}

\item{...}{other arguments to pass to \code{\link[rstan]{sampling}} (e.g., control).}
}
\value{
An object of class \code{stanfit}.
}
\description{
Estimate concentrations (without modeling efficiency) by combining absolute and relative abundance data.
}
\details{
We fit a hierarchical model in Stan to the data, with goal to estimate true concentration for all taxa. There are two available hierarchical models. The first allows for varying efficiency in the relative abundance data, while the second does not allow for varying efficiency (this model). We always recommend using the model that allows for varying efficiency, unless it is known a prior that there is not varying efficiency in the relative abundance data. However, we include this model so that the results of the manuscript can be replicated.
}
\examples{
## load the package, read in example data
library("paramedic")
data(example_16S_data)
data(example_qPCR_data)

## run paramedic (with an extremely small number of iterations, for illustration only)
## on only the first 10 taxa
mod <- no_efficiency(W = example_16S_data[, 1:10], V = example_qPCR_data,
n_iter = 30, n_burnin = 25, n_chains = 1, stan_seed = 4747)

}
\seealso{
\code{\link[rstan]{stan}} and \code{\link[rstan]{sampling}} for specific usage of the \code{stan} and \code{sampling} functions.
}
