% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paramedic-package.R
\docType{package}
\name{paramedic-package}
\alias{paramedic}
\alias{paramedic-package}
\title{paramedic: Cleaner Parameter Documentation}
\description{
The goal of paramedic is to...
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jonthegeek/paramedic}
  \item \url{https://jonthegeek.github.io/paramedic/}
  \item Report bugs at \url{https://github.com/jonthegeek/paramedic/issues}
}

}
\author{
\strong{Maintainer}: Jon Harmon \email{jonthegeek@gmail.com} (\href{https://orcid.org/0000-0003-4781-4346}{ORCID})

}
\keyword{internal}
