% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paramedic.R
\name{paramedic}
\alias{paramedic}
\title{Run a Stan algorithm to estimate concentrations from combined qPCR and br16S data}
\usage{
paramedic(W, V, N = dim(W)[1], q = dim(W)[2], q_obs = dim(V)[2],
  stan_model = "src/stan_files/variable_efficiency.stan",
  n_iter = 10500, n_burnin = 10000, n_chains = 4, stan_seed = 4747,
  params_to_save = c("mu", "Sigma", "beta", "e"), inits_lst = NULL,
  ...)
}
\arguments{
\item{W}{The relative abundance data, e.g., from broad range 16S sequencing with "universal" primers.}

\item{V}{The absolute abundance data, e.g., from taxon-specific qPCR primers.}

\item{N}{Optional. The number of samples. Defaults to the row-length of W.}

\item{q}{Optional. The number of taxa observed with the relative abundance technology. Defaults to the column-length of W.}

\item{q_obs}{Optional. The number of taxa observed with the absolute abundance technology. Defaults to the column-length of V.}

\item{stan_model}{The Stan algorithm to fit to the data. Expects a file path.}

\item{n_iter}{The total number of iterations per chain to be run by the Stan algorithm. Defaults to 10500.}

\item{n_burnin}{The total number of warmups per chain to be run by the Stan algorithm. Defaults to 10000.}

\item{n_chains}{The total number of chains to be run by the Stan algorithm. Defaults to 4.}

\item{stan_seed}{The random number seed to initialize.}

\item{params_to_save}{A character vector of the parameters to save. Defaults to "mu", "Sigma", "beta", "e". TODO(Amy): clarify.}

\item{...}{other arguments to pass to \code{\link[rstan]{stan}}.}
}
\value{
An object of class \code{stanfit}.
}
\description{
Estimate concentrations (and efficiencies, if applicable) using the specified Stan algorithm to combine qPCR and br16S data.
}
\details{
We fit a hierarchical model in Stan to the data, with goal to estimate true concentration for all taxa. There are two available hierarchical models. The two only differ in the way that the true concentration is parameterized. While these hierarchical models are mathematically identical, they are fit differently by Stan. The first hierarchical model is a "centered" model, where \deqn{\beta ~ N(0, \sqrt{50})} \deqn{\log \Sigma ~ N(0, \sqrt{50})} \deqn{\log \mu ~ N(\beta, \Sigma).} We call this model \code{variable_efficiency_centered.stan}. The second hierarchical model is a "noncentered" model, where \deqn{\beta ~ N(0, \sqrt{50})} \deqn{\log \Sigma ~ N(0, \sqrt{50})} \deqn{\log \gamma ~ N(0, I)} \deqn{\log \mu = \sqrt{\Sigma} \log \gamma + \beta.} We call this model \code{variable_efficiency.stan}.
}
\examples{
## load the package, read in example data
library("paramedic")
data(simple_example_data)

## process the example data
q <- 3
q_obs <- 2
processed_data <- process_data(full_data = simple_example_data, br_inds = 1:q,
qpcr_inds = (q + 1):(q + 1 + q_obs),
pcr_plus_br_inds = 1:q_obs,
regex_thr = NA, regex_cps = "", llod = 0,
m_min = 1000, div_num = 1000)

## run paramedic (with an extremely small number of iterations, for illustration only)
mod <- paramedic(W = processed_data$br, V = processed_data$qpcr, q = q, q_obs = q_obs,
stan_model = "src/stan_files/variable_efficiency.stan", n_iter = 30, n_burnin = 25, n_chains = 1, stan_seed = 4747,
params_to_save = c("mu", "Sigma", "beta", "e"))

}
\seealso{
\code{\link[rstan]{stan}} for specific usage of the \code{stan} function.
}
