% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print-and-summary.R
\name{print.paramedic}
\alias{print.paramedic}
\title{Print method for paramedic objects}
\usage{
\method{print}{paramedic}(x, ...)
}
\arguments{
\item{x}{The fitted object}

\item{...}{Ignored}
}
\value{
Returns \code{x}, invisibly.
}
\description{
The \code{print} method for paramedic objects displays a summary of the 
parameter estimates from the fitted model.
}
