% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_paramedic_centered.R
\name{run_paramedic_centered}
\alias{run_paramedic_centered}
\title{Run a Stan algorithm to estimate concentrations from combined absolute and relative abundance data}
\usage{
run_paramedic_centered(W, V, n_iter = 10500, n_burnin = 10000,
  n_chains = 4, stan_seed = 4747, inits_lst = NULL, ...)
}
\arguments{
\item{W}{The relative abundance data, e.g., from broad range 16S sequencing with "universal" primers.}

\item{V}{The absolute abundance data, e.g., from taxon-specific absolute primers.}

\item{n_iter}{The total number of iterations per chain to be run by the Stan algorithm. Defaults to 10500.}

\item{n_burnin}{The total number of warmups per chain to be run by the Stan algorithm. Defaults to 10000.}

\item{n_chains}{The total number of chains to be run by the Stan algorithm. Defaults to 4.}

\item{stan_seed}{The random number seed to initialize.}

\item{inits_lst}{An optional list of initial values of the parameters. Must be a named list; see \code{\link[rstan]{stan}}.}

\item{...}{other arguments to pass to \code{\link[rstan]{sampling}} (e.g., control).}
}
\value{
An object of class \code{stanfit}.
}
\description{
Estimate concentrations (and efficiencies, if applicable) by combining absolute and relative abundance data.
}
\details{
We fit a hierarchical model in Stan to the data, with goal to estimate true concentration for all taxa. There are two available hierarchical models. The two only differ in the way that the true concentration is parameterized. While these hierarchical models are mathematically identical, they are fit differently by Stan. The first hierarchical model is a "centered" model, where \deqn{\beta ~ N(0, \sqrt{50})} \deqn{\log \Sigma ~ N(0, \sqrt{50})} \deqn{\log \mu ~ N(\beta, \Sigma).} We call this model \code{variable_efficiency_centered.stan}. The second hierarchical model is a "noncentered" model, where \deqn{\beta ~ N(0, \sqrt{50})} \deqn{\log \Sigma ~ N(0, \sqrt{50})} \deqn{\log \gamma ~ N(0, I)} \deqn{\log \mu = \sqrt{\Sigma} \log \gamma + \beta.} We call this model \code{variable_efficiency.stan}.
In most cases, we suggest using the noncentered model. However, if the model does not converge in a reasonable amount of time using the noncentered model, consider trying the centered model.
}
\examples{
## load the package, read in example data
library("paramedic")
data(example_16S_data)
data(example_qPCR_data)

## run paramedic (with an extremely small number of iterations, for illustration only)
## on only the first 10 taxa
mod <- run_paramedic_centered(W = example_16S_data[, 1:10], V = example_qPCR_data,
n_iter = 30, n_burnin = 25, n_chains = 1, stan_seed = 4747)

}
\seealso{
\code{\link[rstan]{stan}} and \code{\link[rstan]{sampling}} for specific usage of the \code{stan} and \code{sampling} functions.
}
