% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_example_data.R
\docType{data}
\name{simple_example_data}
\alias{simple_example_data}
\title{Simulated data from relative and absolute abundance views}
\format{
`simple_example_data`: a matrix with 10 rows and 5 columns, 
where the first three columns are relative abundance data and the final two columns are absolute abundance data.
}
\usage{
simple_example_data
}
\description{
A simple simulated dataset, to show how we might process prior to running \code{run_paramedic}.
}
\keyword{datasets}
