% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print-and-summary.R
\name{summary.paramedic}
\alias{summary.paramedic}
\title{Summary method for paramedic objects}
\usage{
summary.paramedic(object, ...)
}
\arguments{
\item{object}{The fitted object}

\item{...}{Other arguments to \code{rstan::summary}}
}
\value{
Returns a summary of \code{x}.
}
\description{
The \code{summary} method for paramedic objects displays a summary of the 
parameter estimates from the fitted model.
}
