% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_predict.R
\name{posterior_predict.paramedic}
\alias{posterior_predict.paramedic}
\alias{posterior_predict}
\title{Draw from the posterior predictive distribution}
\usage{
\method{posterior_predict}{paramedic}(
  object,
  W = NULL,
  V = NULL,
  X = V[, 1, drop = FALSE],
  draws = NULL,
  alpha_sigma = 2,
  kappa_sigma = 1,
  alpha_phi = 0,
  beta_phi = 0,
  k = 0,
  sigma_xi = 1,
  ...
)
}
\arguments{
\item{object}{An object of class \code{"paramedic"}, 
resulting from a call to \code{run_paramedic}.}

\item{W}{The new relative abundance data, e.g., 
from broad range 16S sequencing with "universal" primers. 
Expects data (e.g., matrix, data.frame, tibble) with sample identifiers 
in the first column. Sample identifiers must be the same between W and V, 
and the column must have the same name in W and V.}

\item{V}{The new absolute abundance data, e.g., from taxon-specific 
absolute primers. Expects data (e.g., matrix, data.frame, tibble) with 
sample identifiers in the first column. Sample identifiers must be the 
same between W and V, and the column must have the same name in W and V.}

\item{X}{The new covariate data. Expects data (e.g., matrix, data.frame, 
tibble) with sample identifiers in the first column. Sample identifiers
 must be the same between W, V, and X, and the column must have the same 
 name in W, V, and X. If X only consists of the subject identifiers, 
 then no covariates are used.}

\item{draws}{the number of draws to return. The default and maximum number 
of draws is the size of the posterior sample.}

\item{alpha_sigma}{Hyperparameter specifying the shape parameter of the 
prior distribution on \eqn{\sigma_e}. Defaults to 2.}

\item{kappa_sigma}{Hyperparameter specifying the scale parameter of the 
prior distribution on \eqn{\sigma_e}. Defaults to 1.}

\item{alpha_phi}{Hyperparameter specifying the shape parameter of the 
prior distribution on \eqn{\phi}. Defaults to 0; a negative binomial 
model can be specified if both \code{alpha_phi} and \code{beta_phi} are 
nonzero.}

\item{beta_phi}{Hyperparameter specifying the rate parameter of the prior 
distribution on \eqn{\phi}. Defaults to 0; a negative binomial model can 
be specified if both \code{alpha_phi} and \code{beta_phi} are nonzero.}

\item{k}{the number of batches that the relative abundance data W were 
analyzed in. If k = 0 (the default), then batch effects are not considered. 
(currently not used)}

\item{sigma_xi}{Hyperparameters specifying the variance of efficiencies 
over batches. Only used if \code{k} is greater than zero. 
Defaults to 1. (currently not used)}

\item{...}{Ignored}
}
\value{
A list of \code{draws} by \code{ncol(W)} matrices (for taxon-level 
  parameters) or \code{draws} by \code{nrow(W)} by \code{ncol(W)} arrays 
  (for individual-level parameters and data) of simulations from the 
  posterior predictive distribution. Each row of the matrices, and each 
  of the first dimension of the arrays, denotes the predictions generated 
  using a single draw of the model parameters from the posterior distribution.
}
\description{
Predict concentrations (and efficiencies, if applicable), 
absolute abundances, and relative abundances based on the posterior 
distributions from a previously fitted model resulting from a call to 
\code{run_paramedic}.
}
